Finally displaying an image
Don't laugh too hard. Baby steps.
